The notices are fixed
